requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.2
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
Werkzeug==3.0.3
